src/requirements.txt
google-ai-generativelanguage==0.6.10
google-generativeai
PyGitHub
github3.py==1.3.0
unidiff
wcmatch
